integrated strapi in homepage at some more places and fixed ui of homeoage, fixed blog pagination, cretaed new page for construction with strapi integrated
