+++ src/app/(group1)/construction/page.js
@@ -1,125 +1,55 @@
-"use client";
+import Feedback from '@/components/section/feedback'
+import ProductCard from '@/components/ui/cards/productCard'
+import SectionTitle from '@/components/ui/sectionTitle'
 
-import { useEffect, useState } from "react";
-import Feedback from "@/components/section/feedback";
-import ConstructionAllCard from "@/components/ui/cards/constructionAllCard";
-import SectionTitle from "@/components/ui/sectionTitle";
+export const metadata = {
+    title: "Kalpana Struct-Con -- Product Archive",
+    description: "Kalpana Struct-Con is a next js and tailwind css website",
+};
 
-// A small helper to generate slugs consistently
-function generateSlug(str = "") {
-  return str
-    .toLowerCase()
-    .replace(/[^a-z0-9\s-]/g, "") // Remove invalid chars
-    .replace(/\s+/g, "-")        // Replace spaces with hyphens
-    .replace(/-+/g, "-");        // Remove extra hyphens
+async function getConstructionData() {
+    const res = await fetch("https://kscplcms.cubeone.in/api/construction-main-page?populate[mainPage][populate]=*", { cache: "no-store" });
+    if (!res.ok) throw new Error("Failed to fetch construction data");
+    return res.json();
 }
 
-export default function All() {
-  const [projects, setProjects] = useState([]);
-  const [error, setError] = useState("");
+const ProductArchive = async () => {
+    const apiData = await getConstructionData();
+    const { mainPage } = apiData.data;
+  
+    // Build productData dynamically using API mainPage fields
+    const productData = [
+        { id: 1, product_name: "Commercial", product_img: mainPage.commercial.url, width: 800, height: 600 },
+        { id: 2, product_name: "Infrastructure", product_img: mainPage.infrastructure.url, width: 800, height: 600 },
+        { id: 3, product_name: "Institutional", product_img: mainPage.institutional.url, width: 800, height: 600 },
+        { id: 4, product_name: "Residential and Commercial", product_img: mainPage.residentialAndCommercial.url, width: 800, height: 600 },
+        { id: 5, product_name: "Residential", product_img: mainPage.residential.url, width: 800, height: 600 },
+        { id: 6, product_name: "Show all", product_img: mainPage.showAll.url, width: 800, height: 600 },
+    ];
 
-  useEffect(() => {
-    async function fetchProjects() {
-      try {
-        const res = await fetch(
-          "https://kscplcms.cubeone.in/api/projects?populate=projectDetails.MainImage&populate=projectDetails.Images"
-        );
-
-        // Check if the response is okay
-        if (!res.ok) {
-          throw new Error(`Failed to fetch: ${res.status}`);
-        }
-
-        const data = await res.json();
-
-        // If the data structure is as you showed, we can safely check data.data
-        if (!data || !data.data) {
-          throw new Error("Invalid data structure received from API");
-        }
-
-        // Map the API response to the required format
-        const formattedData = data.data.map((item) => {
-          const details = item.projectDetails || {};
-          const {
-            MainImage,
-            Title = "No Title",
-            Client = "No Client",
-            Category = "",
-          } = details;
-
-          // Generate slug from the title
-          const slug = generateSlug(Title);
-
-          return {
-            id: item.id,
-            img: MainImage?.url || "",
-            width: MainImage?.width || 800,
-            height: MainImage?.height || 600,
-            name: Title,
-            position: Client,
-            slug,
-            category: Category,
-          };
-        });
-
-        setProjects(formattedData);
-      } catch (err) {
-        console.error("Error fetching projects:", err);
-        setError(err.message || "Error fetching projects");
-      }
-    }
-
-    fetchProjects();
-  }, []);
-
-  // Optionally handle and show any error
-  if (error) {
     return (
-      <div className="container pt-20">
-        <p className="text-red-600">Failed to load projects: {error}</p>
-      </div>
-    );
-  }
-
-  return (
-    <>
-      <section>
-        <div className="container-fluid">
-          <SectionTitle
-            sectionName="Projects"
-            sectionTitle="Construction"
-            sectionDesc="Where Imagination Takes Flight, and Excellence Blossoms"
-          />
-        </div>
-        <div className="container lg:pt-30 2sm:pt-20 pt-14">
-          <div className="grid lg:grid-cols-3 2sm:grid-cols-2 gap-7">
-            {projects.map(
-              ({
-                id,
-                img,
-                width,
-                height,
-                name,
-                position,
-                slug,
-                category,
-              }) => (
-                <ConstructionAllCard
-                  key={id}
-                  img={img}
-                  width={width}
-                  height={height}
-                  name={name}
-                  position={position}
-                  slug={slug}
-                  category={category}
-                />
-              )
-            )}
-          </div>
-        </div>
-      </section>
-      <Feedback />
-    </>
-  );
+        <>
+            <section>
+                <div className='container-fluid'>
+                    <SectionTitle 
+                      sectionName={mainPage.backgroundTitle} 
+                      sectionTitle={mainPage.title} 
+                      sectionDesc={mainPage.subtitle}
+                    />
+                </div>
+                <div className='container mt-30'>
+                    <div className='grid lg:grid-cols-3 2sm:grid-cols-2 gap-x-7 gap-y-17.5 mb-25'>
+                        {
+                            productData.map(({ id, product_img, product_name, width, height }) => 
+                                <ProductCard key={id} id={id} product_img={product_img} product_name={product_name} width={width} height={height} />
+                            )
+                        }
+                    </div>
+                </div>
+            </section>
+            <Feedback />
+        </>
+    )
 }
+
+export default ProductArchive;
